api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Niccolo Salvini <47143429+NiccoloSalvini@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
sqlalchemy = "^2.0.41"
pydantic-extra-types = "^2.10.4"
uvicorn = "^0.34.2"
pydantic = {extras = ["email"], version = "^2.11.4"}
python-dotenv = "^1.1.0"
python-jose = "^3.4.0"
passlib = "^1.7.4"
python-multipart = "^0.0.20"
bcrypt = "^4.3.0"
pandas = "^2.2.3"
psycopg2-binary = "^2.9.10"
httpx = "^0.28.1"
fastapi-admin = "^1.0.4"
sqladmin = "^0.20.1"
jinja2 = "^3.1.6"
itsdangerous = "^2.2.0"
loguru = "^0.7.3"
jwt = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
